+++ changelog.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## v2.0.0
+
+#### Breaking Changes
+
+- The ability to send single strings into the main function (call the individual functions)
+- The ability to send partial objects to the main function
+
+#### New
+
+- `number` which can be used to validate a credit card number
+- `cvn` which can be used to validate a credit card cvn/security code
+- `expired` which can be used to validate a credit cards expiration date
+- `matches` which can be used to see if a cvn and a card number match (the cvn length matches the card type)
+- `validation` which is the core functionality and only accepts an object
+- An info property to the results that says if it failed by a rule or if it failed matching
+- Type Errors have been added when the correct type of value isn't provided (String or Number)
+
+#### Fixes
+
+- Improved the overall structure of the code which allowed removal of a lot of internals
+- Cleaner organization now that the module isn't supporting so many different types
+
 ## v1.4.0
 
 - Added: Defaults for most validation methods